requirements.txt
# The Moon Drip BAR - Roasting Cost Calculator
# 더문 드립바 로스팅 원가 계산기 의존성

# Core
streamlit==1.38.0
pandas==2.2.3
numpy<2  # NumPy 1.x (opencv-python 호환성)

# Database
sqlalchemy==2.0.44
pysqlite3-binary

# Visualization
plotly==5.24.1

# Utilities
openpyxl==3.1.5
xlsxwriter==3.2.0

# Report & Export
reportlab==4.0.9
python-dateutil==2.8.2

# Image Processing & OCR
pytesseract==0.3.10
# easyocr==1.7.0  # Claude API로 대체
opencv-python==4.8.1.78
pdf2image==1.16.3
python-Levenshtein==0.25.0
dateparser==1.2.0
Pillow>=8.0.0

# Claude API for OCR
anthropic>=0.73.0
google-generativeai>=0.4.0

# Scheduling (자동화)
APScheduler==3.10.4

# Configuration
python-dotenv>=1.0.0

# Authentication
passlib==1.7.4
bcrypt==4.1.2
